// types/report/geographicDataType.ts
export type TGeographicRow = {
  division: string;
  city: string;
  thana: string;
  totalOrders: number;
  deliveredOrders: number;
  returnedOrders: number;
  totalRevenue: number;
  deliverySuccessRate: string;
  topCourierService: string;
  averageOrderValue: string;
  returnRate: string;
  topCourierCount: number;
};